refactor: share numeric helper
