+++ src/main/java/magma/app/StatementParser.java
@@ -226,33 +226,13 @@ class StatementParser {
             return rest.isEmpty() ? "unknown" : rest;
         }
 
-        if (isNumeric(trimmed)) return "number";
+        if (ParseHelper.isNumeric(trimmed)) return "number";
         if (trimmed.equals("true") || trimmed.equals("false")) return "boolean";
         if (trimmed.length() >= 2 && trimmed.startsWith("\"") && trimmed.endsWith("\"")) {
             return "string";
         }
 
         return "unknown";
-    }
-
-    private static boolean isNumeric(String s) {
-        if (s.isEmpty()) return false;
-        var i = 0;
-        if (s.charAt(0) == '-') {
-            if (s.length() == 1) return false;
-            i = 1;
-        }
-        var dot = false;
-        for (; i < s.length(); i++) {
-            var c = s.charAt(i);
-            if (c == '.') {
-                if (dot) return false;
-                dot = true;
-                continue;
-            }
-            if (c < '0' || c > '9') return false;
-        }
-        return true;
     }
 
     static java.util.Map<String, String> paramVars(String tsParams) {
